Refactor to seperate draggable and droppable logic.
