+++ frontend/src/components/control/ResumioApp/InputApp/InputApp.jsx
@@ -1,18 +1,20 @@
 import React from 'react'
 import { AppSection } from '../../../style/ResumioApp/ResumioAppStyles'
 import { InputForm, 
-  SectionList,
   NewDraftBttn, 
   DraftTitle} from '../../../style/ResumioApp/InputApp/InputAppStyles'
-import ResumioSection from './ResumioSection'
+import SectionList from './SectionList'
 import { useForm, FormProvider } from 'react-hook-form'
-import { DragDropContext, Droppable } from "react-beautiful-dnd"
+import { DragDropContext,  } from "react-beautiful-dnd"
 
 
 export default function InputApp({currentDraft, setCurrentDraft}) {
   const formMethods = useForm()
   const onSubmit = data => console.log(data)
 
+  function onDragEnd(result){
+    console.log(result)
+  }
   //check if currentDraft is empty
   if (Object.keys(currentDraft).length === 0){
     return <h2>Please select a draft</h2>
@@ -25,29 +27,8 @@ export default function InputApp({currentDraft, setCurrentDraft}) {
         <DraftTitle>{currentDraft.title}</DraftTitle>
         <FormProvider {...formMethods}>
           <InputForm onSubmit={formMethods.handleSubmit(onSubmit)}>
-            <DragDropContext>
-              <Droppable droppableId='inputForm'>
-                {(provided, snapshot) => (
-                  <SectionList 
-                    ref={provided.innerRef} 
-                    {...provided.droppableProps}
-                    style={{ backgroundColor: snapshot.isDraggingOver ? 'blue' : 'grey' }}
-                  >
-                      {currentDraft.resume.map((section, index) => {
-                        const sectionType = Object.keys(section)[0]
-                        return (
-                          <ResumioSection 
-                            key={sectionType}
-                            sectionType={sectionType} 
-                            section={section} 
-                            resumeIndex={index} 
-                          />
-                        )
-                      })}
-                      {provided.placeholder}
-                  </SectionList>
-                )}
-              </Droppable>
+            <DragDropContext onDragEnd={onDragEnd}>
+              <SectionList currentDraft={currentDraft} />
             </DragDropContext>
             <button type='submit'>Submit</button>
           </InputForm>
